requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
apscheduler==3.10.4
Pillow==10.4.0
piexif==1.1.3
aiofiles==23.2.1
